refactor: rename filename from create_event.py to event_crud.py.
Before adding other crud functions, rename the file name for better clarity.
